+++ Tic-Tac-Toe/InvalidTurnException.cs
@@ -3,14 +3,14 @@ namespace TicTacToe;
 [Serializable]
 public class InvalidTurnException : Exception
 {
-    public static string GetMsg(int cell, int player, uint player1, uint player2, uint state)
+    public static string GetMsg(int cell, int player, ulong player1, ulong player2, ulong state)
     {
         return $"Invalid {player} move: {cell}. {player1} vs {player2}. {state}";
     }
 
-    public InvalidTurnException() : base(GetMsg(-1, -1, unchecked ((uint)-1), unchecked ((uint)-1), unchecked ((uint)-1))) { }
-    public InvalidTurnException(int cell, int player, uint player1, uint player2, uint state) : base(GetMsg(cell, player, player1, player2, state)) { }
-    public InvalidTurnException(int cell, int player, uint player1, uint player2, uint state, Exception inner) : base(GetMsg(cell, player, player1, player2, state), inner) { }
+    public InvalidTurnException() : base(GetMsg(-1, -1, unchecked ((ulong)-1), unchecked ((ulong)-1), unchecked ((ulong)-1))) { }
+    public InvalidTurnException(int cell, int player, ulong player1, ulong player2, ulong state) : base(GetMsg(cell, player, player1, player2, state)) { }
+    public InvalidTurnException(int cell, int player, ulong player1, ulong player2, ulong state, Exception inner) : base(GetMsg(cell, player, player1, player2, state), inner) { }
     [Obsolete("Obsolete .ctor")]
     protected InvalidTurnException(
         System.Runtime.Serialization.SerializationInfo info,
